vl-map-layer en vl-map-features-layer nog wat beter scheiden
