+++ test/e2e/components/vl-map-draw-point-action.js
@@ -1,7 +1,7 @@
 const VlMapAction = require('./vl-map-action');
 
 class VlMapDrawPointAction extends VlMapAction {
-  async draw({x = 10, y = 10} = {x: 10, y: 10}) {
+  async draw({x, y} = {x: 10, y: 10}) {
     const map = await this.getMap();
     await map.scrollIntoView();
     const actions = this.driver.actions();
