+++ test/e2e/pages/vl-map-draw-actions.page.js
@@ -8,9 +8,6 @@ class VlMapDrawActionsPage extends Page {
   async getMapWithDrawPointAction() {
     return new VlMap(this.driver, '#map-with-draw-point-action');
   }
-  async getMapWithDrawLineAction() {
-    return new VlMap(this.driver, '#map-with-draw-line-action');
-  }
 
   async getMapWithDrawLineAction() {
     return new VlMap(this.driver, '#map-with-draw-line-action');
