+++ test/e2e/pages/vl-map-select-action.page.js
@@ -1,8 +1,7 @@
 const VlMapSelectAction = require('../components/vl-map-select-action');
 const VlMap = require('../components/vl-map');
-const VlMapLayer = require('../components/vl-map-layer');
 const {Page, Config} = require('vl-ui-core').Test;
-const {By} = require('vl-ui-core').Test.Setup;
+const {By, assert} = require('vl-ui-core').Test.Setup;
 
 class VlMapSelectActionPage extends Page {
   async getSelectAction() {
@@ -14,8 +13,10 @@ class VlMapSelectActionPage extends Page {
   }
 
   async clickPointFeature(id) {
-    const layer = await this._getLayer('#map-layer');
     const map = await this._getMap('#map-with-select-action');
+    const layers = await map.getLayers();
+    assert.isNotEmpty(layers);
+    const layer = layers[0];
     await layer.clickPointFeatureOnMap(id, map);
     return this._waitForFeatureToBeSelected(id);
   }
@@ -41,8 +42,8 @@ class VlMapSelectActionPage extends Page {
     return new VlMapSelectAction(this.driver, selector);
   }
 
-  async _getLayer(selector) {
-    return new VlMapLayer(this.driver, selector);
+  async _getLayer(layerClass, selector) {
+    return new layerClass(this.driver, selector);
   }
 
   async _getMap(selector) {
